feat(Calendar): 캘린더 사용자 초대

+++ src/main/java/com/fluffy/SharingCalendar/user/repository/UserRepository.java
@@ -1,13 +1,14 @@
 package com.fluffy.SharingCalendar.user.repository;
 
 import com.fluffy.SharingCalendar.user.domain.User;
+import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     boolean existsByLoginId(String loginId);
     Optional<User> findByLoginId(String loginId);
 }
